More the EnsureTrailingSlash code More the EnsureTrailingSlash code out of GitHubSettingsReader Into a new Abstractions / helper class as it's not specific to github urls
